% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvtu_functions.R
\name{compute_scores}
\alias{compute_scores}
\title{Title}
\usage{
compute_scores(x, EXPR, var = "SCALE", FUN = sum, offset = 0, ...)
}
\arguments{
\item{x}{}

\item{EXPR}{}

\item{var}{}

\item{FUN}{}

\item{offset}{}

\item{...}{}
}
\description{
Title
}
