% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoi.R
\name{hoi}
\alias{hoi}
\title{Title}
\usage{
hoi(x, mud = hoi_mudel)
}
\arguments{
\item{x}{}

\item{mud}{Mudel (vaikimisi kasutatakse RVTU 1. laine andmete põhjal koostatud GRM mudelit, kuid võib ette anda ka mõne teise GRM mudeli)}
}
\value{
faktorskoorid
}
\description{
Title
}
