% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvtu_functions.R
\name{regularize_mk}
\alias{regularize_mk}
\title{Title}
\usage{
regularize_mk(x, direction = "long", keep.na = TRUE)
}
\arguments{
\item{x}{}

\item{direction}{}

\item{keep.na}{}
}
\description{
Title
}
