% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvtu_functions.R
\name{split_levels}
\alias{split_levels}
\title{Title}
\usage{
split_levels(x, sep1 = "|", sep2 = ":", num = TRUE)
}
\arguments{
\item{x}{}

\item{sep1}{}

\item{sep2}{}

\item{num}{}
}
\description{
Title
}
