% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvtu_functions.R
\name{use_labels.data.frame}
\alias{use_labels.data.frame}
\title{Title}
\usage{
\method{use_labels}{data.frame}(x, vn, as.factor = FALSE)
}
\arguments{
\item{x}{}

\item{vn}{}

\item{as.factor}{}
}
\description{
Title
}
