% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoi.R
\name{vthoi}
\alias{vthoi}
\title{Vaimse tervisega seotud heaolu indeks (VTHOI)}
\usage{
vthoi(x, mud = hoi_mudel, ...)
}
\arguments{
\item{x}{andmestik (8 tunnust õigest järjekorras: POS1..POS3, B1...B4, Tervis_eh)}

\item{mud}{Mudel (vaikimisi kasutatakse RVTU 1. laine andmete põhjal koostatud GRM mudelit, kuid võib ette anda ka mõne teise GRM mudeli)}
}
\value{
faktorskooride vektor (puuduvate andmetega ridades on faktorskoori asemele kirjutatud NA)
}
\description{
Vaimse tervisega seotud heaolu indeks (VTHOI)
}
